+++ python/pyclipr/module.cpp
@@ -26,7 +26,6 @@ typedef Eigen::Matrix<double,Eigen::Dynamic,1>   EigenVec1d;
 typedef Eigen::Matrix<double,Eigen::Dynamic,2>   EigenVec2d;
 typedef Eigen::Matrix<double,Eigen::Dynamic,3>   EigenVec3d;
 
-
 static void myZCB(const Clipper2Lib::Point64& e1bot, const Clipper2Lib::Point64& e1top,
                   const Clipper2Lib::Point64& e2bot, const Clipper2Lib::Point64& e2top,
                   Clipper2Lib::Point64& pt) {
@@ -52,22 +51,21 @@ static void myZCB(const Clipper2Lib::Point64& e1bot, const Clipper2Lib::Point64&
     pt.z = maxZ;
 };
 
+pyclipr::EigenVec2d path2EigenVec2d(const Clipper2Lib::Path64 &path, double scaleFactor) {
 
+    pyclipr::EigenVec2d eigPath(path.size(), 2);
 
-Clipper2Lib::Path64 simplifyPath(const Clipper2Lib::Path64 &path, double epsilon, bool isOpenPath = false)
-{
-    return Clipper2Lib::SimplifyPath(path, epsilon, isOpenPath);
+    for (uint64_t i=0; i<path.size(); i++) {
+        eigPath(i,0) = path[i].x;
+        eigPath(i,1) = path[i].y;
+    }
+    eigPath *= scaleFactor;
+
+    return eigPath;
 }
 
-Clipper2Lib::Paths64 simplifyPaths(const Clipper2Lib::Paths64 &paths, double epsilon, bool isOpenPath = false)
+Clipper2Lib::Path64 createPath(const py::array_t<double> &path, double scaleFactor)
 {
-    return Clipper2Lib::SimplifyPaths(paths, epsilon, isOpenPath);
-}
-
-bool orientation(const py::array_t<double> &path, const float scaleFactor = 1000)
-{
-
-    Clipper2Lib::Path64 p;
 
     if (path.ndim() != 2)
         throw std::runtime_error("Number of dimensions must be two");
@@ -75,10 +73,12 @@ bool orientation(const py::array_t<double> &path, const float scaleFactor = 1000
     if (!(path.shape(1) == 2 || path.shape(1) == 3))
         throw std::runtime_error("Path must be nx2, or nx3");
 
+    auto r = path.unchecked<2>();
+
+    Clipper2Lib::Path64 p;
+
     // Resize the path list
     p.reserve(path.shape(0));
-
-    auto r = path.unchecked<2>();
 
     if(path.shape(1) == 2) {
         for(uint64_t i=0; i < path.shape(0); i++)
@@ -89,29 +89,122 @@ bool orientation(const py::array_t<double> &path, const float scaleFactor = 1000
             p.push_back(Clipper2Lib::Point64(r(i,0) * scaleFactor, r(i,1) * scaleFactor, r(i,2)));
 
     }
+    return p;
+}
+
+std::vector<Clipper2Lib::Path64> createPaths(const std::vector<py::array_t<double>> &paths, double scaleFactor)  {
+
+    /* Check all array of paths have a consistent size and dimension */
+
+    int numDims = paths[0].ndim();
+    int numCols = paths[0].shape(1);
+
+    for (auto path: paths) {
+        if (path.ndim() != numDims)
+            throw std::runtime_error("All paths must have same numpy dimensions");
+
+        if (path.shape(1) != numCols)
+            throw std::runtime_error("All paths must have same number of dimensions");
+    }
+
+    std::vector<Clipper2Lib::Path64> clipperPaths;
+    for (auto path: paths) {
+        const Clipper2Lib::Path64 & cPath = createPath(path, scaleFactor);
+        clipperPaths.push_back(cPath);
+    }
+
+    return clipperPaths;
+}
+
+py::object simplifyPath(const py::array_t<double> &path, double epsilon, double scaleFactor = 1e3, bool isOpenPath = false)
+{
+    if (scaleFactor < std::numeric_limits<double>::epsilon()) {
+        throw std::runtime_error("Scale factor cannot be zero");
+    }
+
+    Clipper2Lib::Path64 clipperPath = createPath(path, scaleFactor);
+    Clipper2Lib::Path64 simplifiedPath = Clipper2Lib::SimplifyPath(clipperPath, epsilon, isOpenPath);
+
+    // Convert this back into an eigen Vector2D
+    EigenVec2d simpPathOut = path2EigenVec2d(simplifiedPath, 1.0 / scaleFactor);
+    return py::cast(simpPathOut);
+}
+
+py::object simplifyPaths(std::vector<py::array_t<double>> &paths, double epsilon, double scaleFactor = 1e3, bool isOpenPath = false)
+{
+    if (scaleFactor < std::numeric_limits<double>::epsilon()) {
+        throw std::runtime_error("Scale factor cannot be zero");
+    }
+
+    std::vector<Clipper2Lib::Path64> clipperPaths = createPaths(paths, scaleFactor);
+    std::vector<Clipper2Lib::Path64> simpPaths =  Clipper2Lib::SimplifyPaths(clipperPaths, epsilon, isOpenPath);
+
+    std::vector<EigenVec2d> simpPathsOut;
+    for (auto simpPath : simpPaths) {
+
+        // Convert this back into an eigen Vector2D
+        EigenVec2d simpPathOut = path2EigenVec2d(simpPath, 1.0 / scaleFactor);
+        simpPathsOut.push_back(simpPathOut);
+    }
+
+    return py::cast(simpPathsOut);
+}
+
+bool orientation(const py::array_t<double> &path, const double scaleFactor = 1e3)
+{
+    if (scaleFactor < std::numeric_limits<double>::epsilon()) {
+        throw std::runtime_error("Scale factor cannot be zero");
+    }
+
+    Clipper2Lib::Path64 p = createPath(path, scaleFactor);
 
     return Clipper2Lib::IsPositive(p);
 }
 
-Clipper2Lib::Paths64 polyTreeToPaths64(const Clipper2Lib::PolyTree64 &polytree)
+py::object polyTreeToPaths64(const Clipper2Lib::PolyTree64 &polytree, double scaleFactor = 1.0)
 {
-    return Clipper2Lib::PolyTreeToPaths64(polytree);
+    auto paths = Clipper2Lib::PolyTreeToPaths64(polytree);
+
+    std::vector<EigenVec2d> closedOut;
+
+    for (auto &path : paths) {
+        auto eigPath = path2EigenVec2d(path, scaleFactor);
+        closedOut.push_back(eigPath);
+    }
+
+    return py::cast(closedOut);
 }
 
-Clipper2Lib::PathsD polyTreeToPathsD(const Clipper2Lib::PolyTreeD &polytree)
+py::object polyTreeToPathsD(const Clipper2Lib::PolyPathD &polytree)
 {
-    return Clipper2Lib::PolyTreeToPathsD(polytree);
+    auto paths = Clipper2Lib::PolyTreeToPathsD(polytree);
+
+    std::vector<EigenVec2d> closedOut;
+
+    for (auto &path : paths) {
+
+        EigenVec2d eigPath(path.size(), 2);
+
+        for (uint64_t i=0; i<path.size(); i++) {
+            eigPath(i,0) = double(path[i].x);
+            eigPath(i,1) = double(path[i].y);
+        }
+
+        closedOut.push_back(eigPath);
+    }
+
+    return py::cast(closedOut);
 }
 
 void applyScaleFactor(const Clipper2Lib::PolyPath64 & polyPath,
-                      Clipper2Lib::PolyPathD &newPath, uint64_t scaleFactor, bool invert = true) {
-// Create a recursive copy of the structure
+                      Clipper2Lib::PolyPathD &newPath, double scaleFactor, bool invert = true) {
+    // Create a recursive copy of the structure
 
     for(uint64_t i = 0; i < polyPath.Count(); i++)
     {
         Clipper2Lib::Path64 path = polyPath[i]->Polygon();
         Clipper2Lib::PolyPathD pathD;
-        newPath.SetScale(1.0 / double(scaleFactor));
+        newPath.SetScale(1.0 / scaleFactor);
 
         auto newChild = newPath.AddChild(path);
 
@@ -125,7 +218,8 @@ class Clipper : public Clipper2Lib::Clipper64 {
 
 public:
 
-    Clipper() : Clipper2Lib::Clipper64(), scaleFactor(1000.0) {
+    Clipper() : Clipper2Lib::Clipper64(), scaleFactor(1000.0)
+    {
         this->SetZCallback(myZCB);
     }
     ~Clipper() {}
@@ -134,8 +228,8 @@ public:
 public:
 
    void addPath(const py::array_t<double> &path,
-                 Clipper2Lib::PathType polyType,
-                 bool isOpen)
+                Clipper2Lib::PathType polyType,
+                bool isOpen)
     {
         Clipper2Lib::Path64 p;
 
@@ -173,6 +267,18 @@ public:
 
     void cleanUp() { this->CleanUp(); }
     void clear() { this->Clear(); }
+
+    void setScaleFactor(const double scaleFactor) {
+        if (scaleFactor < std::numeric_limits<double>::epsilon()) {
+            throw std::runtime_error("Scale factor cannot be zero");
+        }
+
+        this->scaleFactor = scaleFactor;
+    }
+    void setPreserveCollinear(bool val) { this->PreserveCollinear(val); }
+
+    double getScaleFactor() const {  return this->scaleFactor; }
+    bool getPreserveCollinear() const { return this->PreserveCollinear(); }
 
     py::object execute(const Clipper2Lib::ClipType clipType, const Clipper2Lib::FillRule fillRule,
                        bool returnOpenPaths = false, bool returnZ = false) {
@@ -224,7 +330,6 @@ public:
                     eigPath(i,1) = double(path[i].y) / double(scaleFactor);
 
                     if(returnZ) {
-                        //std::cout << "OP: assign z" << path[i].z;
                         eigPathZ(i) = path[i].z;
                     }
                 }
@@ -245,12 +350,12 @@ public:
     py::object execute2(const Clipper2Lib::ClipType clipType, const Clipper2Lib::FillRule fillRule,
                         const bool returnOpenPaths = false, const bool returnZ = false) {
 
-        Clipper2Lib::PolyTree64 polytree;
+        Clipper2Lib::PolyPath64 polytree;
         Clipper2Lib::Paths64 openPaths;
 
         this->Execute(clipType, fillRule, polytree, openPaths);
 
-        Clipper2Lib::PolyTreeD *polytreeCpy = new Clipper2Lib::PolyTreeD();
+        Clipper2Lib::PolyPathD *polytreeCpy = new Clipper2Lib::PolyPathD();
 
         applyScaleFactor(polytree, *polytreeCpy, scaleFactor);
 
@@ -289,24 +394,26 @@ public:
 
     }
 
-public:
-    uint64_t scaleFactor;
+protected:
+    double scaleFactor;
 };
-
 
 
 class ClipperOffset : public Clipper2Lib::ClipperOffset {
 
 public:
-    // Consturctor and destructor
-    ClipperOffset() : Clipper2Lib::ClipperOffset(), scaleFactor(1000.0) {}
+
+    ClipperOffset() : Clipper2Lib::ClipperOffset(), scaleFactor(1000.0)
+    {
+
+    }
     ~ClipperOffset() {}
 
 public:
 
     void addPath(const pybind11::array_t<double>& path,
-                 const Clipper2Lib::JoinType joinType,
-                 const Clipper2Lib::EndType endType = Clipper2Lib::EndType::Polygon)
+                            const Clipper2Lib::JoinType joinType,
+                            const Clipper2Lib::EndType endType = Clipper2Lib::EndType::Polygon)
     {
         Clipper2Lib::Path64 p;
 
@@ -332,6 +439,23 @@ public:
         this->AddPath(p, joinType, endType);
 
     }
+
+    void setScaleFactor(const double scaleFactor) {
+        if (scaleFactor < std::numeric_limits<double>::epsilon()) {
+            throw std::runtime_error("Scale factor cannot be zero");
+        }
+
+        this->scaleFactor = scaleFactor;
+    }
+
+    double getScaleFactor() const {  return this->scaleFactor; }
+    double getArcTolerance() const { return this->ArcTolerance(); }
+    double getMiterLimit() const { return this->MiterLimit(); }
+    bool getPreserveCollinear() const { return this->PreserveCollinear(); }
+
+    void setMiterLimit(double val) { this->MiterLimit(val); }
+    void setArcTolerance(double val) { this->ArcTolerance(val); }
+    void setPreserveCollinear(bool val) { this->PreserveCollinear(val); }
 
     void addPaths(const std::vector<pybind11::array_t<double>> paths,
                   const Clipper2Lib::JoinType joinType,
@@ -371,7 +495,7 @@ public:
 
     void clear() { this->Clear(); }
 
-    std::vector<EigenVec2d> execute(const double delta) {
+    py::object  execute(const double delta) {
 
         Clipper2Lib::Paths64 closedPaths;
 
@@ -391,30 +515,29 @@ public:
             closedOut.push_back(eigPath);
         }
 
-
-        return closedOut;
+        return py::cast(closedOut);
     }
 
-    Clipper2Lib::PolyTreeD * execute2(const double delta) {
+    Clipper2Lib::PolyPathD * execute2(const double delta) {
 
-        Clipper2Lib::PolyTree64 polytree;
+        Clipper2Lib::PolyPath64 polytree;
 
         this->Execute(delta * scaleFactor, polytree);
 
-        Clipper2Lib::PolyTreeD *polytreeCpy = new Clipper2Lib::PolyTreeD();
+        Clipper2Lib::PolyPathD *polytreeCpy = new Clipper2Lib::PolyPathD();
 
         applyScaleFactor(polytree, *polytreeCpy, scaleFactor);
 
         return polytreeCpy;
     }
 
-public:
-    uint64_t scaleFactor;
+protected:
+    double scaleFactor;
 };
 
 } // end of namespace pyclipr
 
-PYBIND11_MODULE(pyclipr, m) {
+PYBIND11_MODULE(pyclipr, m, py::mod_gil_not_used()) {
 
     m.doc() = R"pbdoc(
         PyClipr Module
@@ -428,7 +551,6 @@ PYBIND11_MODULE(pyclipr, m) {
 
 	m.attr("clipperVersion")= CLIPPER2_VERSION;
 
-
     py::class_<Clipper2Lib::PolyPath>(m, "PolyPath")
         /*.def(py::init<>()) */
         .def_property_readonly("level", &Clipper2Lib::PolyPath::Level)
@@ -438,7 +560,7 @@ PYBIND11_MODULE(pyclipr, m) {
         /* .def(py::init<>()) */
         .def_property_readonly("isHole", &Clipper2Lib::PolyPath64::IsHole)
         .def_property_readonly("area",   &Clipper2Lib::PolyPath64::Area)
-        .def_property_readonly("attributes", [](const Clipper2Lib::PolyPath64 &s ) {
+        .def_property_readonly("attributes", [](const Clipper2Lib::PolyPath64 &s ) -> py::object {
             Clipper2Lib::Path64 path = s.Polygon();
 
             pyclipr::EigenVec1d eigPath(path.size(), 1);
@@ -446,9 +568,9 @@ PYBIND11_MODULE(pyclipr, m) {
             for (uint64_t i=0; i<path.size(); i++)
                 eigPath(i,0) = path[i].z;
 
-            return eigPath;
+            return py::cast(eigPath);
         })
-        .def_property_readonly("polygon", [](const Clipper2Lib::PolyPath64 &s ) {
+        .def_property_readonly("polygon", [](const Clipper2Lib::PolyPath64 &s ) -> py::object {
             Clipper2Lib::Path64 path = s.Polygon();
 
             pyclipr::EigenVec2d eigPath(path.size(), 3);
@@ -458,7 +580,7 @@ PYBIND11_MODULE(pyclipr, m) {
                 eigPath(i,1) = path[i].y;
                 //eigPath(i,2) = path[i].z;
             }
-            return eigPath;
+            return py::cast(eigPath);
         })
         .def_property_readonly("children", [](const Clipper2Lib::PolyPath64 &s ) {
             std::vector<const Clipper2Lib::PolyPath64 *> children;
@@ -476,7 +598,7 @@ PYBIND11_MODULE(pyclipr, m) {
             /* .def(py::init<>()) */
             .def_property_readonly("isHole", &Clipper2Lib::PolyPathD::IsHole)
             .def_property_readonly("area",   &Clipper2Lib::PolyPathD::Area)
-            .def_property_readonly("attributes", [](const Clipper2Lib::PolyPathD &s ) {
+            .def_property_readonly("attributes", [](const Clipper2Lib::PolyPathD &s) -> py::object {
                 Clipper2Lib::PathD path = s.Polygon();
 
                 pyclipr::EigenVec1d eigPath(path.size(), 1);
@@ -484,9 +606,9 @@ PYBIND11_MODULE(pyclipr, m) {
                 for (uint64_t i=0; i<path.size(); i++)
                     eigPath(i,0) = path[i].z;
 
-                return eigPath;
+                return py::cast(eigPath);
             })
-            .def_property_readonly("polygon", [](const Clipper2Lib::PolyPathD &s ) {
+            .def_property_readonly("polygon", [](const Clipper2Lib::PolyPathD &s ) ->  py::object {
                 Clipper2Lib::PathD path = s.Polygon();
 
                 pyclipr::EigenVec2d eigPath(path.size(), 2);
@@ -494,9 +616,8 @@ PYBIND11_MODULE(pyclipr, m) {
                 for (uint64_t i=0; i<path.size(); i++) {
                     eigPath(i,0) = path[i].x;
                     eigPath(i,1) = path[i].y;
-                    //eigPath(i,2) = path[i].z;
                 }
-                return eigPath;
+                return py::cast(eigPath);
             })
             .def_property_readonly("children", [](const Clipper2Lib::PolyPathD &s ) {
                 std::vector<const Clipper2Lib::PolyPathD *> children;
@@ -506,13 +627,12 @@ PYBIND11_MODULE(pyclipr, m) {
                 return children;
             })
             .def_property_readonly("count", &Clipper2Lib::PolyPathD::Count)
-            .def("__len__", [](const Clipper2Lib::PolyTreeD &s ) { return s.Count(); });
+            .def("__len__", [](const Clipper2Lib::PolyPathD &s ) { return s.Count(); });
 
 
     m.def("polyTreeToPaths64", &pyclipr::polyTreeToPaths64, py::return_value_policy::automatic)
-    .def("orientation", &pyclipr::orientation, py::arg("path"), py::arg("scaleFactor") = 1000,
-         py::return_value_policy::automatic, R"(
-        This function returns the orientation of a path. Orientation will return `true` if the polygon's orientation
+     .def("orientation", &pyclipr::orientation, py::arg("path"),  py::arg("scaleFactor") = 1000, py::return_value_policy::automatic, R"(
+        This function returns the orientation of a path. Orientation will return `True` if the polygon's orientation
         is counter-clockwise.
 
         :param path: A 2D numpy array of shape (n, 2) or (n, 3) where n is the number of vertices in the path.
@@ -520,64 +640,71 @@ PYBIND11_MODULE(pyclipr, m) {
         :return: `True` if the polygon's orientation is counter-clockwise, `False` otherwise.
         )" )
     .def("polyTreeToPaths", &pyclipr::polyTreeToPaths64, py::return_value_policy::automatic)
-    .def("simplifyPath", &pyclipr::simplifyPath, py::arg("path"), py::arg("epsilon"), py::arg("isOpenPath") = false,
+    .def("simplifyPath", &pyclipr::simplifyPath, py::arg("path"), py::arg("epsilon"), py::arg("scaleFactor")  = 1000, py::arg("isOpenPath") = false,
                         py::return_value_policy::automatic, R"(
             This function removes vertices that are less than the specified epsilon distance from an imaginary line
-            that passes through its 2 adjacent vertices. Logically, smaller epsilon values will be less aggressive
+            that passes through its two adjacent vertices. Logically, smaller epsilon values will be less aggressive
             in removing vertices than larger epsilon values.
 
             :param path: A 2D numpy array of shape (n, 2) or (n, 3) where n is the number of vertices in the path.
-            :param epsilon: The maximum distance a vertex can be from an imaginary line that passes through its 2 adjacent vertices.
-            :param isOpenPath: If `True``, the path is treated as an open path. If `False`, the path is treated as a closed path.
-            :return: Simplified path)"
+            :param epsilon: The maximum distance a vertex can be from an imaginary line that passes through its two adjacent vertices.
+            :param scaleFactor: The scaleFactor applied to the path during simplification
+            :param isOpenPath: If `True`, the path is treated as an open path. If `False`, the path is treated as a closed path.
+            :return: Simplified path
+            )"
      )
-    .def("simplifyPaths", &pyclipr::simplifyPaths, py::arg("paths"), py::arg("epsilon"), py::arg("isOpenPath") = false,
-                        py::return_value_policy::automatic, R"(
+    .def("simplifyPaths", &pyclipr::simplifyPaths, py::arg("paths"), py::arg("epsilon"),
+                                                   py::arg("scaleFactor")  = 1000, py::arg("isOpenPath") = false,
+                                                   py::return_value_policy::automatic, R"(
             This function removes vertices that are less than the specified epsilon distance from an imaginary line
-            that passes through its 2 adjacent vertices. Logically, smaller epsilon values will be less aggressive
+            that passes through its two adjacent vertices. Logically, smaller epsilon values will be less aggressive
             in removing vertices than larger epsilon values.
 
             :param paths: A list of 2D points (x,y) that define the path. Tuple or a numpy array may be provided for the path
             :param epsilon: The maximum distance a vertex can be from an imaginary line that passes through its 2 adjacent vertices.
-            :param isOpenPath: If `True``, the path is treated as an open path. If `False`, the path is treated as a closed path.
+            :param scaleFactor: The scaleFactor applied to the path during simplification
+            :param isOpenPath: If `True`, the path is treated as an open path. If `False`, the path is treated as a closed path.
             :return: None
             )"
     );
 
 
-    py::class_<pyclipr::Clipper> clipper(m, "Clipper");
+    py::class_<pyclipr::Clipper> clipper(m, "Clipper",  R"(
+    The Clipper class manages the process of clipping polygons using a number of different Boolean operations,
+    by providing a list of open or closed subject and clipping paths. These are internally represented with Int64
+    precision, that requires the user to specify a scaleFactor. )"
+    );
 
 
     /*
      * Path types are exported for convenience because these tend to be used often
      */
-    py::enum_<Clipper2Lib::PathType>(m, "PathType")
-        .value("Subject", Clipper2Lib::PathType::Subject)
-        .value("Clip",    Clipper2Lib::PathType::Clip)
-    .export_values();
+    py::enum_<Clipper2Lib::PathType>(m, "PathType",  "The path type")
+        .value("Subject", Clipper2Lib::PathType::Subject, "The subject path")
+        .value("Clip",    Clipper2Lib::PathType::Clip, "The clipping path")
+        .export_values();
 
     /*
      * Boolean ops are exported for convenience because these tend to be used often
      */
-    py::enum_<Clipper2Lib::ClipType>(m, "ClipType")
-        .value("Union",        Clipper2Lib::ClipType::Union)
-        .value("Difference",   Clipper2Lib::ClipType::Difference)
-        .value("Intersection", Clipper2Lib::ClipType::Intersection)
-        .value("Xor",          Clipper2Lib::ClipType::Xor)
-    .export_values();
+    py::enum_<Clipper2Lib::ClipType>(m, "ClipType", py::arithmetic(), "The clipping operation type")
+        .value("Union",        Clipper2Lib::ClipType::Union, "Union operation")
+        .value("Difference",   Clipper2Lib::ClipType::Difference, "Difference operation")
+        .value("Intersection", Clipper2Lib::ClipType::Intersection, "Intersection operation")
+        .value("Xor",          Clipper2Lib::ClipType::Xor, "XOR operation")
+        .export_values();
 
-    py::enum_<Clipper2Lib::FillRule>(m, "FillRule")
-        .value("EvenOdd",  Clipper2Lib::FillRule::EvenOdd)
-        .value("NonZero",  Clipper2Lib::FillRule::NonZero)
-        .value("Positive", Clipper2Lib::FillRule::Positive)
-        .value("Negative", Clipper2Lib::FillRule::Negative);
+    py::enum_<Clipper2Lib::FillRule>(m, "FillRule", py::arithmetic(), "The fill rule to be used for the clipping operation")
+        .value("EvenOdd",  Clipper2Lib::FillRule::EvenOdd, "Even and Odd Fill")
+        .value("NonZero",  Clipper2Lib::FillRule::NonZero, "Non-Zero Fill")
+        .value("Positive", Clipper2Lib::FillRule::Positive, "Positive Fill")
+        .value("Negative", Clipper2Lib::FillRule::Negative, "Negative Fill");
 
     clipper.def(py::init<>())
-        .def_readwrite("scaleFactor", &pyclipr::Clipper::scaleFactor, R"(
-            Scale factor to be for transforming the input and output vectors. The default is 1000 )"
+        .def_property("scaleFactor", &pyclipr::Clipper::getScaleFactor, &pyclipr::Clipper::setScaleFactor, R"(
+            The scale factor to be for transforming the input and output vectors. The default is 1000. )"
          )
-        .def_property("preserveCollinear", [](const pyclipr::Clipper &s ) { return s.PreserveCollinear();}
-                                         , [](pyclipr::Clipper &s, bool val ) { return s.PreserveCollinear(val);}, R"(
+        .def_property("preserveCollinear", &pyclipr::Clipper::getPreserveCollinear, &pyclipr::Clipper::setPreserveCollinear, R"(
              By default, when three or more vertices are collinear in input polygons (subject or clip),
              the Clipper object removes the 'inner' vertices before clipping. When enabled the PreserveCollinear property
              prevents this default behavior to allow these inner vertices to appear in the solution.
@@ -596,15 +723,14 @@ PYBIND11_MODULE(pyclipr, m) {
             :param path: A list paths, each consisting 2D points (x,y) that define the path. A Tuple or a numpy array may be provided
             :param pathType: A PathType enum value that indicates whether the path is a subject or a clip path.
             :param isOpen: A boolean value that indicates whether the path is closed or not. Default is `False`
-            :param returnZ: If `True`, returns a seperate array of the Z attributes for clipped paths. Default is `False`
+            :param returnZ: If `True`, returns a separate array of the Z attributes for clipped paths. Default is `False`
             :return: None
         )" )
-        .def("execute", &pyclipr::Clipper::execute,
-                          py::arg("clipType"),
-                          py::arg("fillRule") =  Clipper2Lib::FillRule::EvenOdd,
-                          py::kw_only(), py::arg("returnOpenPaths") = false,
-                          py::arg("returnZ") = false,
-                          py::return_value_policy::take_ownership, R"(
+        .def("execute", &pyclipr::Clipper::execute,  py::arg("clipType"),
+                                                     py::arg_v("fillRule", Clipper2Lib::FillRule::EvenOdd, "FIllRule.EvenOdd"),
+                                                     py::kw_only(), py::arg("returnOpenPaths") = false,
+                                                     py::arg("returnZ") = false,
+                                                     py::return_value_policy::take_ownership, R"(
             The execute method performs the Boolean clipping operation on the polygons or paths that have been added
             to the clipper object. This method will return a list of paths from the result. The default fillRule is
             even-odd typically used for the representation of polygons.
@@ -612,15 +738,14 @@ PYBIND11_MODULE(pyclipr, m) {
             :param clipType: The ClipType or the clipping operation to be used for the paths
             :param fillRule: A FillType enum value that indicates the fill representation for the paths
             :param returnOpenPaths: If `True`, returns a tuple consisting of both open and closed paths
-            :param returnZ: If `True`, returns a seperate array of the Z attributes for clipped paths. Default is `False`
+            :param returnZ: If `True`, returns a separate array of the Z attributes for clipped paths. Default is `False`
             :return: A resultant paths that have been clipped )"
             )
-        .def("execute2", &pyclipr::Clipper::execute2,
-                          py::arg("clipType"),
-                          py::arg("fillRule") =  Clipper2Lib::FillRule::EvenOdd,
-                          py::kw_only(), py::arg("returnOpenPaths") = false,
-                          py::arg("returnZ") = false,
-                          py::return_value_policy::take_ownership, R"(
+        .def("execute2", &pyclipr::Clipper::execute2, py::arg("clipType"),
+                                                      py::arg_v("fillRule", Clipper2Lib::FillRule::EvenOdd, "FIllRule.EvenOdd"),
+                                                      py::kw_only(), py::arg("returnOpenPaths") = false,
+                                                      py::arg("returnZ") = false,
+                                                      py::return_value_policy::take_ownership, R"(
             The execute2 method performs the Boolean clipping operation on the polygons or paths that have been added
             to the clipper object. TThis method will return a PolyTree of the result structuring the output into the hierarchy of
             the paths that form the exterior and interior polygon.
@@ -630,54 +755,56 @@ PYBIND11_MODULE(pyclipr, m) {
             :param clipType: The ClipType or the clipping operation to be used for the paths
             :param fillRule: A FillType enum value that indicates the fill representation for the paths
             :param returnOpenPaths: If `True`, returns a tuple consisting of both open and closed paths. Default is `False`
-            :param returnZ: If `True`, returns a seperate array of the Z attributes for clipped paths. Default is `False`
+            :param returnZ: If `True`, returns a separate array of the Z attributes for clipped paths. Default is `False`
             :return: A resultant polytree of the clipped paths )"
             )
 
-        .def("executeTree", &pyclipr::Clipper::execute2,
-                        py::arg("clipType"),
-                        py::arg("fillRule") =  Clipper2Lib::FillRule::EvenOdd,
-                        py::kw_only(), py::arg("returnOpenPaths") = false,
-                        py::arg("returnZ") = false,
-                        py::return_value_policy::take_ownership, R"(
-        The executeTree method performs the Boolean clipping operation on the polygons or paths that have been added
+        .def("executeTree", &pyclipr::Clipper::execute2,  py::arg("clipType"),
+                                                          py::arg_v("fillRule", Clipper2Lib::FillRule::EvenOdd, "FIllRule.EvenOdd"),
+                                                          py::kw_only(), py::arg("returnOpenPaths") = false,
+                                                          py::arg("returnZ") = false,
+                                                          py::return_value_policy::take_ownership, R"(
+
+        The `executeTree` method performs the Boolean clipping operation on the polygons or paths that have been added
         to the clipper object. TThis method will return a PolyTree of the result structuring the output into the hierarchy of
         the paths that form the exterior and interior polygon.
 
-        The default fillRule is even-odd typically used for the representation of polygons.
+        The default `FillRule` is even-odd typically used for the representation of polygons.
 
         :param clipType: The ClipType or the clipping operation to be used for the paths
         :param fillRule: A FillType enum value that indicates the fill representation for the paths
         :param returnOpenPaths: If `True`, returns a tuple consisting of both open and closed paths. Default is `False`
-        :param returnZ: If `True`, returns a seperate array of the Z attributes for clipped paths. Default is `False`
+        :param returnZ: If `True`, returns a separate array of the Z attributes for clipped paths. Default is `False`
         :return: A resultant paths that have been clipped )"
         )
         .def("clear", &pyclipr::Clipper::clear, R"(The clear method removes all the paths from the Clipper object.)" )
         .def("cleanUp", &pyclipr::Clipper::cleanUp);
 
 
+    py::class_<pyclipr::ClipperOffset> clipperOffset(m, "ClipperOffset", R"(
+    The ClipperOffset class manages the process of offsetting  (inflating/deflating)
+    both open and closed paths using a number of different join types and end types.
+    The library user will rarely need to access this unit directly since it will generally
+    be easier to use the InflatePaths function when doing polygon offsetting.)"
+     );
 
-    py::class_<pyclipr::ClipperOffset> clipperOffset(m, "ClipperOffset");
+    py::enum_<Clipper2Lib::JoinType>(m, "JoinType", py::arithmetic(), "The join type to be used for the offsetting / inflation of paths")
+        .value("Square", Clipper2Lib::JoinType::Square, "Square join type")
+        .value("Round",  Clipper2Lib::JoinType::Round, "Round join type")
+        .value("Miter",  Clipper2Lib::JoinType::Miter, "Miter join type");
 
-    py::enum_<Clipper2Lib::JoinType>(m, "JoinType")
-        .value("Square", Clipper2Lib::JoinType::Square)
-        .value("Round",  Clipper2Lib::JoinType::Round)
-        .value("Miter",  Clipper2Lib::JoinType::Miter);
-
-    py::enum_<Clipper2Lib::EndType>(m, "EndType")
-        .value("Square",  Clipper2Lib::EndType::Square)
-        .value("Butt",    Clipper2Lib::EndType::Butt)
-        .value("Joined",  Clipper2Lib::EndType::Joined)
-        .value("Polygon", Clipper2Lib::EndType::Polygon)
-        .value("Round",   Clipper2Lib::EndType::Round);
-
+    py::enum_<Clipper2Lib::EndType>(m, "EndType", py::arithmetic(), "The end type to be used for the offsetting / inflation of paths")
+        .value("Square",  Clipper2Lib::EndType::Square, "Square end type")
+        .value("Butt",    Clipper2Lib::EndType::Butt, "Butt end type")
+        .value("Joined",  Clipper2Lib::EndType::Joined, "Joined end type")
+        .value("Polygon", Clipper2Lib::EndType::Polygon, "Polygon end type")
+        .value("Round",   Clipper2Lib::EndType::Round, "Round end type");
 
     clipperOffset.def(py::init<>())
-        .def_readwrite("scaleFactor", &pyclipr::ClipperOffset::scaleFactor, R"(
-            Scale factor to be for transforming the input and output vectors. The default is 1000 )"
-         )
-        .def_property("arcTolerance", [](const Clipper2Lib::ClipperOffset &s ) { return s.ArcTolerance();}
-                                    , [](Clipper2Lib::ClipperOffset &s, double val ) { return s.ArcTolerance(val);}, R"(
+        .def_property("scaleFactor", &pyclipr::ClipperOffset::getScaleFactor, &pyclipr::ClipperOffset::setScaleFactor,
+            R"(Scale factor for transforming the input and output vectors. The default is 1000.)"
+        )
+        .def_property("arcTolerance", &pyclipr::ClipperOffset::getArcTolerance,   &pyclipr::ClipperOffset::setArcTolerance, R"(
             Firstly, this field/property is only relevant when JoinType = Round and/or EndType = Round.
 
             Since flattened paths can never perfectly represent arcs, this field/property specifies a maximum acceptable
@@ -685,69 +812,60 @@ PYBIND11_MODULE(pyclipr, m) {
             'smoothness' up to a point though at a cost of performance and in creating more vertices to construct the arc.
 
             The default ArcTolerance is 0.25 units. This means that the maximum distance the flattened path will deviate
-            from the 'true' arc will be no more than 0.25 units (before rounding). )"
-        )
-        .def_property("miterLimit", [](const Clipper2Lib::ClipperOffset &s ) { return s.MiterLimit();}
-                                  , []( Clipper2Lib::ClipperOffset &s, double val ) { return s.MiterLimit(val);}, R"(
+            from the 'true' arc will be no more than 0.25 units (before rounding). )"  )
+        .def_property("miterLimit", &pyclipr::ClipperOffset::getMiterLimit, &pyclipr::ClipperOffset::setMiterLimit,  R"(
              This property sets the maximum distance in multiples of delta that vertices can be offset from their original
              positions before squaring is applied. (Squaring truncates a miter by 'cutting it off' at 1 x delta distance
              from the original vertex.)
 
              The default value for MiterLimit is 2 (ie twice delta). )"
         )
-        .def_property("preserveCollinear", [](const Clipper2Lib::ClipperOffset &s ) { return s.PreserveCollinear();}
-                                         , [](Clipper2Lib::ClipperOffset &s, bool val ) { return s.PreserveCollinear(val);}, R"(
+        .def_property("preserveCollinear", &pyclipr::ClipperOffset::getPreserveCollinear, &pyclipr::ClipperOffset::setPreserveCollinear, R"(
              By default, when three or more vertices are collinear in input polygons (subject or clip),
-             the Clipper object removes the 'inner' vertices before clipping. When enabled the PreserveCollinear property
+             the Clipper object removes the 'inner' vertices before clipping. When enabled the `PreserveCollinear` property
              prevents this default behavior to allow these inner vertices to appear in the solution. )"
          )
         .def("addPath", &pyclipr::ClipperOffset::addPath, py::arg("path"),
-                         py::arg("joinType"),
-                         py::arg("endType") = Clipper2Lib::EndType::Polygon, R"(
+                                                          py::arg("joinType"),
+                                                          py::arg_v("endType", Clipper2Lib::EndType::Polygon, "Endtype.Polygon"),  R"(
             The addPath method adds one open or closed paths (polygon) to the ClipperOffset object.
 
             :param path: A list of 2D points (x,y) that define the path. Tuple or a numpy array may be provided for the path
-            :param joinType: The JoinType to use for the offseting / inflation of paths
-            :param endType: The EndType to use for the offseting / inflation of paths (default is Polygon)
+            :param joinType: The JoinType to use for the offsetting / inflation of paths
+            :param endType: The EndType to use for the offsetting / inflation of paths (default is Polygon)
             :return: None )"
         )
         .def("addPaths", &pyclipr::ClipperOffset::addPaths, py::arg("path"),
-                          py::arg("joinType"),
-                          py::arg("endType") = Clipper2Lib::EndType::Polygon, R"(
+                                                            py::arg("joinType"),
+                                                            py::arg_v("endType", Clipper2Lib::EndType::Polygon, "Endtype.Polygon"), R"(
             The addPath method adds one or more open / closed paths to the ClipperOffset object.
 
             :param path: A list of paths consisting of 2D points (x,y) that define the path. Tuple or a numpy array may be provided for each path
-            :param joinType: The JoinType to use for the offseting / inflation of paths
-            :param endType: The EndType to use for the offseting / inflation of paths
+            :param joinType: The JoinType to use for the offsetting / inflation of paths
+            :param endType: The EndType to use for the offsetting / inflation of paths
             :return: None)"
         )
-        .def("execute", &pyclipr::ClipperOffset::execute,
-                          py::arg("delta"),
-                          py::return_value_policy::take_ownership, R"(
-            The `execute` method performs the offseting/inflation operation on the polygons or paths that have been added
+        .def("execute", &pyclipr::ClipperOffset::execute,  py::arg("delta"), py::return_value_policy::take_ownership, R"(
+            The `execute` method performs the offsetting/inflation operation on the polygons or paths that have been added
             to the clipper object. This method will return a list of paths from the result.
 
-            :param delta: The offset to apply to the inflation/offseting of paths and segments
+            :param delta: The offset to apply to the inflation/offsetting of paths and segments
             :return: The resultant offset paths
         )")
-        .def("execute2", &pyclipr::ClipperOffset::execute2,
-                          py::arg("delta"),
-                          py::return_value_policy::take_ownership, R"(
-            The `execute` method performs the offseting/inflation operation on the polygons or paths that have been added
+        .def("execute2", &pyclipr::ClipperOffset::execute2,  py::arg("delta"),  py::return_value_policy::take_ownership, R"(
+            The `execute` method performs the offsetting/inflation operation on the polygons or paths that have been added
             to the clipper object. This method will return a PolyTree from the result, that considers the hierarchy of the interior and exterior
             paths of the polygon.
 
-            :param delta: The offset to apply to the inflation/ofsetting
+            :param delta: The offset to apply to the inflation/offsetting
             :return: A resultant offset paths created in a PolyTree64 Object )"
         )
-        .def("executeTree", &pyclipr::ClipperOffset::execute2,
-                          py::arg("delta"),
-                          py::return_value_policy::take_ownership, R"(
-            The `executeTree` method performs the offseting/inflation operation on the polygons or paths that have been added
+        .def("executeTree", &pyclipr::ClipperOffset::execute2,  py::arg("delta"),  py::return_value_policy::take_ownership, R"(
+            The `executeTree` method performs the offsetting/inflation operation on the polygons or paths that have been added
             to the clipper object. This method will return a PolyTree from the result, that considers the hierarchy of the interior and exterior
             paths of the polygon.
 
-            :param delta: The offset to apply to the inflation/ofsetting
+            :param delta: The offset to apply to the inflation/offsetting
             :return: A resultant offset paths created in a PolyTree64 Object )"
          )
         .def("clear", &pyclipr::ClipperOffset::clear, R"(The clear method removes all the paths from the ClipperOffset object.)");